Add argument to override package manager detection
