+++ src/util.mjs
@@ -1,5 +1,5 @@
 import { debuglog } from "node:util";
-import { cwd as processCwd } from "node:process";
+import { cwd as processCwd, env } from "node:process";
 import { basename as baseName, join as pathJoin } from "node:path";
 
 export const log = (msg, ...args) => {
@@ -32,4 +32,18 @@ export const cmdToExecForm = (cmd) => {
     .map((part) => `"${part}"`)
     .join(", ");
   return `[${elements}]`;
+};
+
+export const determinePackageManager = () => {
+  const npmUserAgent = env.npm_config_user_agent;
+  debug("npmUserAgent", npmUserAgent);
+  const npmExecPath = env.npm_execpath;
+  debug("npmExecPath", npmExecPath);
+  const parent = env._;
+  debug("parent", parent);
+
+  if (!npmUserAgent) return "npm";
+  if (npmUserAgent.includes("yarn/")) return "yarn";
+  if (npmUserAgent.includes("pnpm/")) return "pnpm";
+  return "npm";
 };
